Add edit_file support
Following what the Perl API does, the edit_file
method is implemented to fetch an existing and
new URL from the tracker, DAV MOVE from one to
the other and then pass back a mutable file.

Upon closing the file, a length of `-1` is
supplied to avoid having to keep an accurate
internal counter in LargeHTTPFile.